requirements.txt
pysnmp-lextudio>=6.0.0
psutil>=5.9.0
PySide6>=6.5.0
matplotlib>=3.7.0
pytest>=7.4.0
pytest-qt>=4.2.0
